No more p5js
I stopped using p5js since i couldnt decide when to render a frame. This conditioned massively the way i coded sorting functions because i had to write functions for only a step of the sorting proccess. Now with vanilla javascript thats not needed anymore.

For now, the only thing working is the bubbleSort.